Crate pallet_robonomics_liability

The Robonomics runtime module. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use pallet::*;

Modules

economics
Set of approaches to handle economical aspects of agreement.

pallet

signed
Signed liability implementation.

technics
Set of approaches to handle technical aspects of agreement.

traits
Robonomics runtime traits definitions.