Module pallet_robonomics_liability::traits

Robonomics liability traits definitions.

Traits

Agreement

Agreement between two participants around technical/economical aspects.

AgreementProofBuilder

Agreement proof maker.

Processing

Transaction processing of agreement. Usually it consists of balance locking and transfers when liability successfully finished.

RealWorldOracle

Someone who can confirm agreement execution in real world.

Report

During execution one part of agreement should prepare report, it’s parameter of finalize transaction. Technically, lifecycle of liability consist of three steps:

ReportProofBuilder

Report proof maker.