Crate pallet_robonomics_liability

The Robonomics runtime module. This can be compiled with #[no_std], ready for Wasm.

Modules

economics
Set of approaches to handle economical aspects of agreement.

signed
Signed liability implementation.

technics
Set of approaches to handle technical aspects of agreement.

traits
Import module traits. Robonomics runtime traits definitions.

Structs

Module

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

Traits

Trait
Liability module main trait.

Type Definitions

AccountId
Current runtime account identificator.

EconomicalParam
Type synonym for economical trait parameter.

Event
RawEvent specialized for the configuration Trait

LiabilityIndex

ProofParam
Type synonym for liability proof parameter.

TechnicalParam
Type synonym for technical trait parameter.

TechnicalReport
Type synonym for technical report trait parameter.