Trait pallet_robonomics_liability::traits::Agreement

pub trait Agreement<AccountId> {
    type Technical: Parameter;
    type Economical: Parameter;

    fn technical(&self) -> Self::Technical;
    fn economical(&self) -> Self::Economical;
    fn promisee(&self) -> AccountId;
    fn promisor(&self) -> AccountId;
    fn verify(&self) -> bool;
}

Agreement between two participants around technical/economical aspects.

Required Associated Types

type Technical: Parameter

Technical parameter of agreement. Usually in this parameter one party of agreement provide technical information like task description, parameters and some additional data.

type Economical: Parameter

Economical parameter of agreement. Usually in this parameter one party set task cost for another party. To come an agreement both parties should be agree with this parameter.

Required Methods

fn technical(&self) -> Self::Technical

Get techincal parameter of agreement.

fn economical(&self) -> Self::Economical

Get Economical parameter of agreement.

fn promisee(&self) -> AccountId

The client account.

fn promisor(&self) -> AccountId

The executive account.

fn verify(&self) -> bool

Check validity of agreement params proof.