Trait pallet_robonomics_liability::traits::Agreement

pub trait Agreement<T: Technical, E: Economical> {
    type Index: Parameter + BaseArithmetic + Member + Copy + Default;
    type AccountId: Parameter;
    type Proof: Parameter;
    fn new(
        technics: T::Parameter, 
        economics: E::Parameter, 
        promisee: Self::AccountId, 
        promisor: Self::AccountId
    ) -> Self;

    fn check_params(
        &self, 
        proof: &Self::Proof, 
        sender: &Self::AccountId
    ) -> bool;

    fn check_report(
        &self, 
        index: &Self::Index, 
        report: &T::Report, 
        proof: &Self::Proof
    ) -> bool;
}

Agreement between two participants around technical/economical aspects.

Associated Types

type Index: Parameter + BaseArithmetic + Member + Copy + Default

Indexing type.

type AccountId: Parameter

Pariticipant account address.

type Proof: Parameter

Some that could be used as proof of participants agreement.

Loading content...
Required methods

fn new(
technics: T::Parameter,
economics: E::Parameter,
promisee: Self::AccountId,
promisor: Self::AccountId
) -> Self

Create new instance for given technical and economical parameters.

fn check_params(&self, proof: &Self::Proof, sender: &Self::AccountId) -> bool

Check validity of agreement params proof.

fn check_report(
&self,
index: &Self::Index,
report: &T::Report,
proof: &Self::Proof
) -> bool

Check validity of agreement report proof.

Loading content...