Trait pallet_robonomics_liability::traits::Report

pub trait Report<Index, AccountId>: RealWorldOracle {
    type Message: Parameter;

    fn index(&self) -> Index;
    fn sender(&self) -> AccountId;
    fn verify(&self) -> bool;
}

During execution one part of agreement should prepare report, it’s parameter of finalize transaction. Technically, lifecycle of liability consist of three steps:

Pre-open: parties try come to an agreement;

Start: funds of one party locked, second party execute a task;

Finish: funds of one party transfered as same as report of executed task published.

Required Associated Types

type Message: Parameter

The report payload.

Required Methods

fn index(&self) -> Index

Agreement indexing type.

fn sender(&self) -> AccountId

Sender account.

fn verify(&self) -> bool

Check validity report proof.